Move link-related structs to `knowledge::link`
