+++ apps/hash-graph/lib/graph-data/src/knowledge.rs
@@ -1,1 +1,2 @@
 pub mod entity;
+pub mod link;
