+++ apps/hash-graph/lib/graph-data/src/knowledge/entity.rs
@@ -15,6 +15,7 @@ use uuid::Uuid;
 
 use crate::{
     account::AccountId,
+    knowledge::link::LinkData,
     provenance::{OwnedById, ProvenanceMetadata},
 };
 
@@ -39,19 +40,6 @@ impl EntityUuid {
 impl fmt::Display for EntityUuid {
     fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt::Display::fmt(&self.0, fmt)
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[repr(transparent)]
-#[cfg_attr(feature = "postgres", derive(FromSql, ToSql), postgres(transparent))]
-#[cfg_attr(feature = "utoipa", derive(ToSchema))]
-pub struct LinkOrder(i32);
-
-impl LinkOrder {
-    #[must_use]
-    pub const fn new(order: i32) -> Self {
-        Self(order)
     }
 }
 
@@ -98,37 +86,6 @@ impl EntityProperties {
     pub const fn properties(&self) -> &HashMap<BaseUrl, serde_json::Value> {
         &self.0
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[cfg_attr(feature = "utoipa", derive(ToSchema))]
-#[serde(deny_unknown_fields)]
-pub struct EntityLinkOrder {
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        rename = "leftToRightOrder"
-    )]
-    #[cfg_attr(feature = "utoipa", schema(nullable = false))]
-    pub left_to_right: Option<LinkOrder>,
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        rename = "rightToLeftOrder"
-    )]
-    #[cfg_attr(feature = "utoipa", schema(nullable = false))]
-    pub right_to_left: Option<LinkOrder>,
-}
-
-/// The associated information for 'Link' entities
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[cfg_attr(feature = "utoipa", derive(ToSchema))]
-#[serde(deny_unknown_fields, rename_all = "camelCase")]
-pub struct LinkData {
-    pub left_entity_id: EntityId,
-    pub right_entity_id: EntityId,
-    #[serde(flatten)]
-    pub order: EntityLinkOrder,
 }
 
 /// The metadata of an [`Entity`] record.
