+++ apps/hash-graph/tests/integration/postgres/entity.rs
@@ -1,4 +1,4 @@
-use graph_data::knowledge::entity::{EntityLinkOrder, EntityProperties};
+use graph_data::knowledge::{entity::EntityProperties, link::EntityLinkOrder};
 use graph_test_data::{data_type, entity, entity_type, property_type};
 use temporal_versioning::ClosedTemporalBound;
 use type_system::url::{BaseUrl, VersionedUrl};
